Add fallback to numeric term parsing
Currently it is possible to easily mis-configure a field across types easily
especially if they are semantically equivalent. Yet, if type 1 has a size field
mapped as an integer and type 2 as a long faceting will fail since the types can not be
parsed due to how we internally store the types in the index. This patch allows to
be more gentle about parsing and falls back to a 32bit / 64bit parsing.

Closes #3106