Use memory interleave policy 'all' by default if 'numactl' is available
If Control NUMA policy for processes or shared memory is available we
use ' --interleave=all'.
By default, Linux attempts to be smart about memory allocations such
that data is close to the NUMA node on which it runs. For big database
type of applications, this is not the best thing to do if the priority
is to avoid disk I/O. In particular with Lucene, we're heavily
multi-threaded anyway and there is no particular reason to believe that
one NUMA node is "better" than another.

Consequences of allocating unevenly among NUMA nodes can include
excessive page cache eviction when the kernel tries to allocate memory -
such as when restarting the JVM.