Packaging: Add LICENSE and NOTICE files for all core dependencies
Also adds dev-tools/check_license_and_sha.pl which will ensure that the
sha1 file in the licenses/ directory is the same as the sha1 of each
dependency shipped with Elasticsearch, and that each dependency has
a LICENSE file.  Can also be used to update the sha1 file when upgrading
dependencies.