Acquire and hold FS lock while modifying the index.
Today we acquire the FS lock before we modify the index direcotry
ie. during deletes but release it immediately again. We should hold
on to this lock until we are done modifing the content and release afterwards.

This commit introduces a try / with pattern for these scenarios to release the locks again.