Never connect and publish connection during unicast zen ping
UnicastZenPing today establishes real connections to nodes during it's ping
phase that can be used by other parts of the system. Yet, this is potentially dangerous
and undesirable unless the nodes have been fully verified and shoudl be connected to in the
case of a cluster state update or if we join a newly elected master.
For zen-pings this change adds dedicated connections that are only kept alive while pings are running.