Respect `ignore_throttled` in security
We apparently work around many things in the security code when it
gets to resolving wildcards and date expressions. This fixes the
case when an index is throttled and resolved with security.
Down the road we really need to fix this code to ensure we don't miss
changes we make to the `IndexNameExpressionResolver`